refactor: Move the builders up a namespace
The only class in the parent directory, `PhpSpecConfigurationBuilder`,
is also a builder hence I think it makes more sense to have all of them
alongside each other.
