+++ src/PhpSpecAdapterFactory.php
@@ -39,8 +39,8 @@ use function file_get_contents;
 use Infection\AbstractTestFramework\TestFrameworkAdapter;
 use Infection\AbstractTestFramework\TestFrameworkAdapterFactory;
 use Infection\TestFramework\PhpSpec\CommandLine\ArgumentsAndOptionsBuilder;
-use Infection\TestFramework\PhpSpec\Config\Builder\InitialConfigBuilder;
-use Infection\TestFramework\PhpSpec\Config\Builder\MutationConfigBuilder;
+use Infection\TestFramework\PhpSpec\Config\InitialConfigBuilder;
+use Infection\TestFramework\PhpSpec\Config\MutationConfigBuilder;
 use InvalidArgumentException;
 use function method_exists;
 use function sprintf;
