+++ tests/phpunit/PhpSpecAdapterTest.php
@@ -38,8 +38,8 @@ namespace Infection\Tests\TestFramework\PhpSpec;
 use Infection\AbstractTestFramework\Coverage\TestLocation;
 use Infection\TestFramework\PhpSpec\CommandLine\ArgumentsAndOptionsBuilder;
 use Infection\TestFramework\PhpSpec\CommandLineBuilder;
-use Infection\TestFramework\PhpSpec\Config\Builder\InitialConfigBuilder;
-use Infection\TestFramework\PhpSpec\Config\Builder\MutationConfigBuilder;
+use Infection\TestFramework\PhpSpec\Config\InitialConfigBuilder;
+use Infection\TestFramework\PhpSpec\Config\MutationConfigBuilder;
 use Infection\TestFramework\PhpSpec\PhpSpecAdapter;
 use Infection\TestFramework\PhpSpec\VersionParser;
 use PHPUnit\Framework\Attributes\CoversClass;
